Moving math package from external libarary (#11147)
