+++ go/vt/orchestrator/inst/instance.go
@@ -24,8 +24,9 @@ import (
 	"strings"
 	"time"
 
+	math "vitess.io/vitess/go/vt/orchestrator/util"
+
 	"vitess.io/vitess/go/vt/orchestrator/config"
-	"vitess.io/vitess/go/vt/orchestrator/external/golib/math"
 	"vitess.io/vitess/go/vt/vtctl/reparentutil/promotionrule"
 )
