+++ go/vt/orchestrator/util/math.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package math
+package util
 
 func MinInt(i1, i2 int) int {
 	if i1 < i2 {
@@ -44,27 +44,6 @@ func MaxInt64(i1, i2 int64) int64 {
 	return i2
 }
 
-func MinUInt(i1, i2 uint) uint {
-	if i1 < i2 {
-		return i1
-	}
-	return i2
-}
-
-func MaxUInt(i1, i2 uint) uint {
-	if i1 > i2 {
-		return i1
-	}
-	return i2
-}
-
-func MinUInt64(i1, i2 uint64) uint64 {
-	if i1 < i2 {
-		return i1
-	}
-	return i2
-}
-
 func MaxUInt64(i1, i2 uint64) uint64 {
 	if i1 > i2 {
 		return i1
@@ -79,13 +58,6 @@ func MinString(i1, i2 string) string {
 	return i2
 }
 
-func MaxString(i1, i2 string) string {
-	if i1 > i2 {
-		return i1
-	}
-	return i2
-}
-
 // TernaryString acts like a "? :" C-style ternary operator for strings
 func TernaryString(condition bool, resTrue string, resFalse string) string {
 	if condition {
@@ -94,20 +66,12 @@ func TernaryString(condition bool, resTrue string, resFalse string) string {
 	return resFalse
 }
 
-// TernaryString acts like a "? :" C-style ternary operator for ints
+// TernaryInt acts like a "? :" C-style ternary operator for ints
 func TernaryInt(condition bool, resTrue int, resFalse int) int {
 	if condition {
 		return resTrue
 	}
 	return resFalse
-}
-
-// AbsInt is an ABS function for int type
-func AbsInt(i int) int {
-	if i >= 0 {
-		return i
-	}
-	return -i
 }
 
 // AbsInt64 is an ABS function for int64 type
